fix rendering on multiple desktops on MacOS Sonoma
It appears that Sonoma is using an instance of the view for each screen.

Since all of the data structures were global, this seemed to result
in erratic rendering (or just double speed animation).

Move all of the state into a class and put a separate instance into
each view.  As a side effect, each screen now has a unique rendering
as the random number generators diverge pretty quickly.